simplify code get active validator indexes
